layers: Move SubmitInfoConverter to convert_utils
After the code is moved the compiler started to notice unused
VkQueueFlags queue_flags parameter, so it was removed.
