+++ layers/sync/sync_submit.h
@@ -331,33 +331,6 @@ class QueueSyncState {
     QueueId id_;
 };
 
-// The converter needs to be more complex than simply an array of VkSubmitInfo2 structures.
-// In order to convert from Info->Info2, arrays of VkSemaphoreSubmitInfo and VkCommandBufferSubmitInfo
-// structures must be created for the pWaitSemaphoreInfos, pCommandBufferInfos, and pSignalSemaphoreInfos
-// which comprise the converted VkSubmitInfo information. The created VkSubmitInfo2 structure then references the storage
-// of the arrays, which must have a lifespan longer than the conversion, s.t. the ensuing valdation/record operations
-// can reference them.  The resulting VkSubmitInfo2 is then copied into an additional which takes the place of the pSubmits
-// parameter.
-struct SubmitInfoConverter {
-    struct BatchStore {
-        BatchStore(const VkSubmitInfo &info, VkQueueFlags queue_flags);
-
-        static VkSemaphoreSubmitInfo WaitSemaphore(const VkSubmitInfo &info, uint32_t index);
-        static VkCommandBufferSubmitInfo CommandBuffer(const VkSubmitInfo &info, uint32_t index);
-        static VkSemaphoreSubmitInfo SignalSemaphore(const VkSubmitInfo &info, uint32_t index, VkQueueFlags queue_flags);
-
-        std::vector<VkSemaphoreSubmitInfo> waits;
-        std::vector<VkCommandBufferSubmitInfo> cbs;
-        std::vector<VkSemaphoreSubmitInfo> signals;
-        VkSubmitInfo2 info2;
-    };
-
-    SubmitInfoConverter(uint32_t count, const VkSubmitInfo *infos, VkQueueFlags queue_flags);
-
-    std::vector<BatchStore> info_store;
-    std::vector<VkSubmitInfo2> info2s;
-};
-
 struct QueueSubmitCmdState {
     std::shared_ptr<const QueueSyncState> queue;
     SignaledSemaphoresUpdate signaled_semaphores_update;
